pyproject.toml
[tool.poetry]
name = "curso-fastapi"
version = "0.1.0"
description = ""
authors = ["Augusto Bauer Domingos <augustobauerdomingos@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = { extras = ["standard"], version = "^0.112.1" }
sqlalchemy = "^2.0.32"
pydantic-settings = "^2.4.0"
alembic = "^1.13.2"
pwdlib = { extras = ["argon2"], version = "^0.2.0" }
python-multipart = "^0.0.9"
pyjwt = "^2.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
ruff = "^0.6.0"
httpx = "^0.27.0"

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
run = 'fastapi dev curso_fastapi/app.py'
pre_test = 'task lint'
test = 'pytest -s --cov=curso_fastapi -vv'
post_test = 'coverage html'

[tool.taskipy]
hooks = { pre_test = "pre_test", post_test = "post_test" }

[tool.ruff]
exclude = ['migrations', 'alembic']
line-length = 88
indent-width = 4

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
